move logging to its own mod, use in start commands
