+++ iroh-cli/src/config.rs
@@ -62,119 +62,6 @@ impl ConsolePaths {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(default)]
-pub(crate) struct FileLogging {
-    /// RUST_LOG directive to filter file logs.
-    rust_log: EnvFilter,
-    /// Maximum number of files to keep.
-    max_files: NonZeroUsize,
-    /// How often should a new log file be produced.
-    rotation: Rotation,
-}
-
-impl Default for FileLogging {
-    fn default() -> Self {
-        let filter = EnvFilter::default();
-        Self {
-            rust_log: filter,
-            max_files: NonZeroUsize::new(8).expect("clearly non zero"),
-            rotation: Rotation::default(),
-        }
-    }
-}
-
-impl FileLogging {
-    /// Obtain a layer for tracing compatible with the fmt subscriber.
-    ///
-    /// This layer will:
-    /// - use the defauilt [`tracing_subscriber::fmt::format::Format`].
-    /// - include line numbers.
-    /// - create log files in the `logs` dir inside the given `iroh_data_root`.
-    /// - rotate files every [`Self::rotation`].
-    /// - keep at most [`Self::max_files`] log files.
-    /// - use the filtering defined by [`Self::rust_log`].
-    /// - create log files with the name `iroh-<ROTATION_BASED_NAME>.log` (ex: iroh-2024-02-02.log)
-    pub(crate) fn layer(
-        &self,
-        iroh_data_root: &Path,
-    ) -> Result<(
-        impl Layer<tracing_subscriber::FmtSubscriber>,
-        tracing_appender::non_blocking::WorkerGuard,
-    )> {
-        use tracing_appender::rolling;
-        let FileLogging {
-            rust_log,
-            max_files,
-            rotation,
-        } = self;
-
-        let (file_logger, guard) = {
-            let rotation = match rotation {
-                Rotation::Hourly => rolling::Rotation::HOURLY,
-                Rotation::Daily => rolling::Rotation::DAILY,
-                Rotation::Never => rolling::Rotation::NEVER,
-            };
-            let logs_path = iroh_data_root.join("logs");
-
-            let file_appender = rolling::Builder::new()
-                .rotation(rotation)
-                .max_log_files(max_files.get())
-                .filename_prefix("iroh")
-                .filename_suffix("log")
-                .build(logs_path)?;
-            tracing_appender::non_blocking(file_appender)
-        };
-
-        Ok((
-            tracing_subscriber::fmt::Layer::new()
-                .with_line_number(true)
-                .with_writer(file_logger)
-                .with_filter(rust_log.layer()),
-            guard,
-        ))
-    }
-}
-
-/// Wrapper to obtain a [`tracing_subscriber::EnvFilter`] that satisfies required bounds.
-#[derive(
-    Debug, Clone, PartialEq, Eq, SerializeDisplay, DeserializeFromStr, derive_more::Display,
-)]
-#[display("{_0}")]
-pub(crate) struct EnvFilter(String);
-
-impl FromStr for EnvFilter {
-    type Err = <tracing_subscriber::EnvFilter as FromStr>::Err;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // validate the RUST_LOG statement
-        let _valid_env = tracing_subscriber::EnvFilter::from_str(s)?;
-        Ok(EnvFilter(s.into()))
-    }
-}
-
-impl Default for EnvFilter {
-    fn default() -> Self {
-        Self("debug".into())
-    }
-}
-
-impl EnvFilter {
-    pub(crate) fn layer(&self) -> tracing_subscriber::EnvFilter {
-        tracing_subscriber::EnvFilter::from_str(&self.0).expect("validated RUST_LOG statement")
-    }
-}
-
-/// Hoe often should a new file be created for file logs.
-/// Akin to [`tracing_appender::rolling::Rotation`].
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize, Default)]
-enum Rotation {
-    #[default]
-    Hourly,
-    Daily,
-    Never,
-}
-
 /// The configuration for an iroh node.
 #[derive(PartialEq, Eq, Debug, Deserialize, Serialize, Clone)]
 #[serde(default)]
@@ -185,7 +72,7 @@ pub(crate) struct NodeConfig {
     pub(crate) gc_policy: GcPolicy,
     /// Bind address on which to serve Prometheus metrics
     pub(crate) metrics_addr: Option<SocketAddr>,
-    pub(crate) file_logs: FileLogging,
+    pub(crate) file_logs: super::logging::FileLogging,
 }
 
 impl Default for NodeConfig {
