+++ iroh-cli/src/main.rs
@@ -5,6 +5,7 @@ use clap::Parser;
 
 mod commands;
 mod config;
+mod logging;
 
 use crate::commands::Cli;
 
@@ -25,33 +26,5 @@ fn main() -> Result<()> {
 async fn main_impl() -> Result<()> {
     let data_dir = config::iroh_data_root()?;
     let cli = Cli::parse();
-
-    // TODO(@divma): remove
-    // #[cfg(unix)]
-    // if let Some(log_fd) = cli.log_fd {
-    //     use std::fs::File;
-    //     use std::mem::ManuallyDrop;
-    //     use std::os::unix::io::FromRawFd;
-    //
-    //     // SAFETY: We take ownership but ensure it is never dropped, thus we never close the
-    //     // filedescriptor.  So even if the users chooses 0, 1 or 2 we do not close it,
-    //     // making sure those keep working as expected until process termination.
-    //     let inner = unsafe { ManuallyDrop::new(File::from_raw_fd(log_fd)) };
-    //     let writer = ManuallyDropFile(inner);
-    //     tracing_subscriber::registry()
-    //         .with(
-    //             tracing_subscriber::fmt::layer()
-    //                 .event_format(tracing_subscriber::fmt::format().with_line_number(true))
-    //                 .with_writer(writer),
-    //         )
-    //         .with(EnvFilter::from_default_env())
-    //         .init();
-    //     return cli.run(&data_dir).await;
-    // }
-    //
-    // tracing_subscriber::registry()
-    //     .with(tracing_subscriber::fmt::layer().with_writer(std::io::stderr))
-    //     .with(EnvFilter::from_default_env())
-    //     .init();
     cli.run(&data_dir).await
 }
